+++ Profit/src/main/java/com/bsu/CheckServlet.java
@@ -12,7 +12,7 @@ public class CheckServlet extends HttpServlet {
     @Override
     protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException{
         resp.setContentType("application/json");
-        String response = "{\"sucess\":true}";
+        String response = "{\"success\":true}";
         resp.getWriter().println(response);
     }
 }
